replace all byte members by int, add player ctor, add initGame

+++ app/src/main/java/com/thilo20/machikoro/Card.java
@@ -13,5 +13,5 @@ public class Card {
         BLUE, GREEN, RED, VIOLET, GRAY
     }
 
-    byte cost;
+    int cost;
 }
